[refactor] Separate the test for results_manager to organize the structure
